+++ lib/lightning/usage_tracking/report.ex
@@ -1,12 +1,12 @@
-defmodule Lightning.ImpactTracking.Report do
+defmodule Lightning.UsageTracking.Report do
   @moduledoc """
-  Report submitted to ImpactTracker
+  Report submitted to Usage Tracker
 
   """
   use Ecto.Schema
 
   @primary_key {:id, :binary_id, autogenerate: true}
-  schema "impact_tracking_reports" do
+  schema "usage_tracking_reports" do
     field :data, :map
     field :submitted, :boolean
     field :submitted_at, :utc_datetime_usec